+++ nemo/collections/multimodal/speechllm/modules/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from nemo.collections.multimodal import models, modules
+from nemo.collections.multimodal.speechllm.modules.speechllm_perception import *
